+++ src/components/modals/set-resolution.js
